Add definition of send data method
To prevent PyCharm from complaining about using an unknown method I copied the that is injected into our web app by fusion and added it to a reference folder. This code is not intended to be used but to tell PyCharm or other IDEs how this method is going to look like in the final environment.